+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/OrchestrationModuleConfig.java
@@ -9,7 +9,8 @@ import com.sap.ai.sdk.orchestration.model.MaskingModuleConfig;
 import com.sap.ai.sdk.orchestration.model.MaskingModuleConfigProviders;
 import com.sap.ai.sdk.orchestration.model.OutputFilteringConfig;
 import com.sap.ai.sdk.orchestration.model.PromptTemplatingModuleConfigPrompt;
-import com.sap.ai.sdk.orchestration.model.SAPDocumentTranslation;
+import com.sap.ai.sdk.orchestration.model.SAPDocumentTranslationInput;
+import com.sap.ai.sdk.orchestration.model.SAPDocumentTranslationOutput;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Objects;
@@ -96,9 +97,9 @@ public class OrchestrationModuleConfig {
    */
   @Nullable GroundingModuleConfig groundingConfig;
 
-  @Nullable SAPDocumentTranslation inputTranslationConfig;
+  @Nullable SAPDocumentTranslationInput inputTranslationConfig;
 
-  @Nullable SAPDocumentTranslation outputTranslationConfig;
+  @Nullable SAPDocumentTranslationOutput outputTranslationConfig;
 
   /**
    * Creates a new configuration with the given LLM configuration.